requirements.txt
pandas==2.1.0
numpy==1.24.3
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
